requirements.txt
Django>=4.2,<5.0
openai>=0.27.0
pdfplumber>=0.5.28
python-docx>=0.8.11
python-dotenv>=1.0.0
gunicorn>=20.1.0
markdown>=3.4.0